@@ -151,16 +150,16 @@
     "@jsenv/node-module-import-map": "13.4.1",
     "@jsenv/package-publish": "1.5.2",
     "@jsenv/prettier-check-project": "5.6.1",
-    "@jsenv/pwa": "3.0.6",
-    "@vue/component-compiler": "4.2.3",
-    "eslint": "7.23.0",
-    "playwright-chromium": "1.10.0",
-    "playwright-firefox": "1.10.0",
-    "playwright-webkit": "1.10.0",
-    "prettier": "2.2.1",
+    "@jsenv/pwa": "3.0.7",
+    "@vue/component-compiler": "4.2.4",
+    "eslint": "7.26.0",
+    "playwright-chromium": "1.11.0",
+    "playwright-firefox": "1.11.0",
+    "playwright-webkit": "1.11.0",
+    "prettier": "2.3.0",
     "react": "17.0.2",
-    "redux": "4.0.5",
-    "sass": "1.32.8",
+    "redux": "4.1.0",
+    "sass": "1.32.12",
     "vue-template-compiler": "2.6.12"
   }
 }
