@@ -1,6 +1,4 @@
-import { require } from "../require.js"
-
-const { supportsColor } = require("supports-color")
+import { createSupportsColor } from "supports-color"
 
 export const cross = "☓" // "\u2613"
 export const checkmark = "✔" // "\u2714"
@@ -12,7 +10,7 @@ export const green = "\x1b[32m"
 export const grey = "\x1b[39m"
 export const ansiResetSequence = "\x1b[0m"
 
-const processSupportsBasicColor = supportsColor(process.stdout).hasBasic
+const processSupportsBasicColor = createSupportsColor(process.stdout).hasBasic
 
 export const setANSIColor = processSupportsBasicColor
   ? (text, color) => `${color}${text}${ansiResetSequence}`
